Create a module for migration extension methods for LedgerStorage
