+++ src/many-ledger/src/storage.rs
@@ -1,16 +1,15 @@
+pub mod migration_ext;
+
 use crate::error;
 #[cfg(feature = "migrate_blocks")]
 use crate::migration;
-use crate::migration::data::{
-    DataMethods, ACCOUNT_TOTAL_COUNT_INDEX, NON_ZERO_ACCOUNT_TOTAL_COUNT_INDEX,
-};
 use crate::migration::{run_migrations, MigrationMap, MigrationSet};
 use crate::module::validate_account;
 use many_error::ManyError;
 use many_identity::Address;
 use many_modules::abci_backend::AbciCommitInfo;
 use many_modules::account::features::FeatureInfo;
-use many_modules::data::{DataIndex, DataInfo, DataValue};
+use many_modules::data::{DataIndex, DataInfo};
 use many_modules::{account, events, idstore, EmptyReturn};
 use many_protocol::ResponseMessage;
 use many_types::ledger::{Symbol, TokenAmount};
@@ -23,6 +22,7 @@ use std::collections::{BTreeMap, BTreeSet, Bound};
 use std::ops::RangeBounds;
 use std::path::Path;
 use tracing::{debug, info};
+use migration_ext::data::DataExt;
 
 fn _execute_multisig_tx(
     ledger: &mut LedgerStorage,
@@ -1563,63 +1563,5 @@ pub mod tests {
             ))
             .len()
         )
-    }
-}
-
-impl DataMethods for LedgerStorage {
-    fn data_attributes(&self) -> Option<BTreeMap<DataIndex, DataValue>> {
-        self.persistent_store
-            .get(DATA_ATTRIBUTES_KEY)
-            .expect("Error while reading the DB")
-            .map(|x| minicbor::decode(&x).unwrap())
-    }
-
-    fn update_data_attributes(
-        &mut self,
-        from: &Address,
-        to: &Address,
-        amount: TokenAmount,
-        symbol: &Address,
-    ) {
-        if let Some(mut attributes) = self.data_attributes() {
-            let key_to = key_for_account_balance(to, symbol);
-            if self
-                .persistent_store
-                .get(&key_to)
-                .expect("Error communicating with the DB")
-                .is_none()
-            {
-                attributes
-                    .entry(*ACCOUNT_TOTAL_COUNT_INDEX)
-                    .and_modify(|x| {
-                        if let DataValue::Counter(count) = x {
-                            *count += 1;
-                        }
-                    });
-                attributes
-                    .entry(*NON_ZERO_ACCOUNT_TOTAL_COUNT_INDEX)
-                    .and_modify(|x| {
-                        if let DataValue::Counter(count) = x {
-                            *count += 1;
-                        }
-                    });
-            }
-            let balance_from = self.get_balance(from, symbol);
-            if balance_from == amount {
-                attributes
-                    .entry(*NON_ZERO_ACCOUNT_TOTAL_COUNT_INDEX)
-                    .and_modify(|x| {
-                        if let DataValue::Counter(count) = x {
-                            *count -= 1;
-                        }
-                    });
-            }
-            self.persistent_store
-                .apply(&[(
-                    DATA_ATTRIBUTES_KEY.to_vec(),
-                    Op::Put(minicbor::to_vec(attributes).unwrap()),
-                )])
-                .unwrap();
-        }
     }
 }
